[feature] Break roomallocation class down
